code cleanup

+++ Src/MoneyFox.BusinessLogic/FileStore/FileStoreIoBase.cs
@@ -17,11 +17,8 @@ namespace MoneyFox.BusinessLogic.FileStore
 
         public override Stream OpenRead(string path)
         {
-            var fullPath = AppendPath(path);
-            if (!File.Exists(fullPath))
-            {
-                return null;
-            }
+            string fullPath = AppendPath(path);
+            if (!File.Exists(fullPath)) return null;
 
             return File.Open(fullPath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
         }
@@ -30,8 +27,8 @@ namespace MoneyFox.BusinessLogic.FileStore
         {
             try
             {
-                var fullFrom = AppendPath(from);
-                var fullTo = AppendPath(destination);
+                string fullFrom = AppendPath(from);
+                string fullTo = AppendPath(destination);
 
                 if (!File.Exists(fullFrom))
                 {
@@ -42,17 +39,14 @@ namespace MoneyFox.BusinessLogic.FileStore
                 if (File.Exists(fullTo))
                 {
                     if (overwrite)
-                    {
                         File.Delete(fullTo);
-                    } else
-                    {
+                    else
                         return false;
-                    }
                 }
 
                 File.Move(fullFrom, fullTo);
                 return true;
-            } 
+            }
             catch (Exception ex)
             {
                 logger.Error(ex.ToString);
@@ -62,19 +56,18 @@ namespace MoneyFox.BusinessLogic.FileStore
 
         protected override void WriteFileCommon(string path, Action<Stream> streamAction)
         {
-            var fullPath = AppendPath(path);
-            if (File.Exists(fullPath))
-            {
-                File.Delete(fullPath);
-            }
+            string fullPath = AppendPath(path);
+            if (File.Exists(fullPath)) File.Delete(fullPath);
 
-            using (var fileStream = File.OpenWrite(fullPath))
+            using (FileStream fileStream = File.OpenWrite(fullPath))
             {
                 streamAction?.Invoke(fileStream);
             }
         }
-        
+
         protected virtual string AppendPath(string path)
-            => Path.Combine(BasePath, path);
+        {
+            return Path.Combine(BasePath, path);
+        }
     }
 }
